requirements.txt
# python -m pip install -r .\requirements.txt --upgrade
colorama==0.4.6
discord.py==2.4.0
pillow==11.0.0
spotipy==2.24.0
yt-dlp==2024.11.18
PyNaCl==1.5.0
